refactor(servers): extract WindowServerCommandService
